requirements.txt
accelerate==1.0.1
aenum==3.1.16
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
alembic==1.15.2
annotated-types==0.7.0
anthropic==0.62.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrdict==2.0.1
attrs==24.2.0
av==13.1.0
babel==2.16.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
bleach==6.2.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cohere==5.20.0
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.0.2
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
epiweeks==2.3.0
executing==2.1.0
fastapi==0.115.6
fastavro==1.12.1
fastjsonschema==2.20.0
ffmpy==0.5.0
filelock==3.13.1
flash-attn==2.7.0.post2
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.2.0
ftfy==6.3.1
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.40.3
google-auth-httplib2==0.2.1
google-genai==1.29.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grad-cam==1.5.4
gradio==5.9.1
gradio_client==1.5.2
greenlet==3.2.1
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.14.0
hf-xet==1.1.7
httpcore==1.0.6
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
humanize==4.12.3
idna==3.10
imageio==2.36.1
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
monai==1.5.0
monai-generative==0.2.3
monai-weekly==1.6.dev2529
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open-clip-torch==2.23.0
openai==1.78.0
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
optuna==4.3.0
orjson==3.10.13
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.2
peft==0.13.2
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
plotly==5.24.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qwen-vl-utils==0.0.8
referencing==0.35.1
regex==2024.11.6
replicate==1.0.7
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.21.0
rsa==4.9.1
ruff==0.8.6
sacremoses==0.1.1
safehttpx==0.1.6
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
sentry-sdk==2.19.0
setproctitle==1.3.4
setuptools==75.1.0
shellingham==1.5.4
simsimd==6.4.7
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spaces==0.31.1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.5
stringzilla==3.12.5
sympy==1.13.1
tenacity==9.0.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.11.0
timm==1.0.24
tinycss2==1.4.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==0.10.3
torchvision==0.20.1+cu121
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.51.3
triton==3.1.0
trl==0.12.1
ttach==0.0.3
typer==0.15.1
types-python-dateutil==2.9.0.20241003
types-requests==2.32.4.20250913
typing_extensions==4.14.0
tzdata==2024.2
umap-learn==0.5.7
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.2.3
uvicorn==0.34.0
wandb==0.18.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wheel==0.44.0
widgetsnbextension==4.0.13
xai-sdk==1.0.1
xxhash==3.5.0
yarl==1.18.0
